Initialize default weather forecast and process immediately
Added initialization of `this.weatherForecast` with a default value from `this.weatherForecastDataService.getDefaultWeatherForecast()`. Also, called `this.processDayDetailForecast()` right after setting up the subscription to ensure the component has a default forecast and processes it immediately. This helps in setting up initial states and handling cases where the subscription might not emit data immediately.